+++ tests/browser/value.test.js
@@ -13,7 +13,7 @@ describe("Value", function() {
         browser = await Wendigo.createBrowser({log:true});
     });
 
-    beforeEach(async() => {
+    beforeEach(async () => {
         await browser.open(configUrls.forms);
     });
 
@@ -45,21 +45,21 @@ describe("Value", function() {
         assert.strictEqual(value, "default value");
     });
 
-    it("Set Value", async() => {
+    it("Set Value", async () => {
         const changed = await browser.setValue(".input1", "my-val");
         const value = await browser.value(".input1");
         assert.strictEqual(value, "my-val");
         assert.strictEqual(changed, 1);
     });
 
-    it("Set Value From Node", async() => {
+    it("Set Value From Node", async () => {
         const element = await browser.query(".input1");
         await browser.setValue(element, "my-val");
         const value = await browser.value(".input1");
         assert.strictEqual(value, "my-val");
     });
 
-    it("Set Value Multiple Elements", async() => {
+    it("Set Value Multiple Elements", async () => {
         const changed = await browser.setValue("form input", "my-val");
         const value1 = await browser.value(".input1");
         const value2 = await browser.value(".input2");
@@ -68,7 +68,7 @@ describe("Value", function() {
         assert.strictEqual(changed, 2);
     });
 
-    it("Set Value From Non-existing Element", async() => {
+    it("Set Value From Non-existing Element", async () => {
         await utils.assertThrowsAsync(async () => {
             await browser.setValue(".not-element", "my-val");
         }, `Error: Element ".not-element" not found when trying to set value "my-val".`);